+++ reproc/include/reproc/reproc.h
@@ -114,10 +114,10 @@ typedef struct reproc_options {
   Stop actions that are passed to `reproc_stop` in `reproc_destroy` to stop the
   child process.
 
-  When `stop_actions` is 3x `REPROC_STOP_NOOP`, `reproc_destroy` will default to
+  When `stop` is 3x `REPROC_STOP_NOOP`, `reproc_destroy` will default to
   waiting indefinitely for the child process to exit.
   */
-  reproc_stop_actions stop_actions;
+  reproc_stop_actions stop;
 } reproc_options;
 
 /*! Allocate a new `reproc_t` instance on the heap. */
@@ -278,8 +278,7 @@ function, its exit status is returned.
 Actionable errors:
 - `REPROC_ETIMEDOUT`
 */
-REPROC_EXPORT int reproc_stop(reproc_t *process,
-                              reproc_stop_actions stop_actions);
+REPROC_EXPORT int reproc_stop(reproc_t *process, reproc_stop_actions stop);
 
 /*!
 Release all resources associated with `process` including the memory allocated